requirements.txt
numpy==1.22.3
pandas==1.4.2
Pillow==9.3.0
plotly==5.13.0
streamlit==1.20.0
tensorflow==2.10.0
